perf: Intel BTS: Add "splice" output mode
This is a set of modifications to handle program tracing
with Branch Trace Store (BTS) more efficiently and reduce
overhead by approximately 2x.

The original perf behavior was to set up one 16-page buffer per CPU,
and to copy its whole content to a larger place in RAM (128 pages per
core) every time an overflow interrupt or a context switch occurs.
Then, the user-space tracing daemon would copy the contents of this
intermediate buffer to disk.  The trace file being open without the
O_SYNC flag, writing to disk is not synchronized and data is possibly
copied once more to a temporary buffer.

To avoid the time-wasting multiple copy in RAM (especially the one
performed during the handling of an interrupt), we re-implemented this
section to use ring-buffers.  BTS is then configured to store entries
in a sub-buffer; when it is full, BTS is reconfigured to use the next
sub-buffer in the ring-buffer.  That way, there is no urgent need to
copy the BTS data to make room: this data is saved to disk later by
a kernel task.

The size of the sub-buffers is the same as in perf (16 pages), their
number was chosen to allocate a bit less resources than perf does in
its default implementation, that is 16 kernel-space pages plus 128
user-space pages per core (576 KiB per core for original perf, 512 KiB
in our implementation).